--- __mocks__/bounty.ts
@@ -1,11 +1,0 @@
-
-export const mockTrailer = {
-  summary: "trailer summary",
-  description: "trailer descriptio",
-  video: "trailer video",
-  duration: 4,
-  info: {
-    items: ["item 1", "item 2"],
-    title: "info title",
-  },
-};
