+++ __tests__/components/sections/submissions/_partials/SubmissionCard.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { renderWithRedux } from "../../../../../__mocks__/renderWithRedux";
 import { screen } from "@testing-library/react";
 import SubmissionCard from "@/components/sections/submissions/_partials/SubmissionCard";
-import { submission } from "../../../../../__mocks__/challenge";
+import { submission } from "../../../../../__mocks__/fixtures/challenge";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -13,40 +13,37 @@ jest.mock("next/router", () => ({
 }));
 
 describe("Submission Card", () => {
+  beforeEach(()=>{
+    renderWithRedux(<SubmissionCard submission={submission} />);
+  })
   it("should render submission card", () => {
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
     const submissionCard = screen.getByTestId("submissionCardId");
     expect(submissionCard).toBeInTheDocument();
   });
   it("should display the user's display name", () => {
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
-    const displayName = screen.getByText(submission().user.displayName);
+    const displayName = screen.getByText(submission.user.displayName);
     expect(displayName).toBeInTheDocument();
   });
 
   it("should display the submission text", () => {
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
-    const submissionText = screen.getByText(submission().text);
+    const submissionText = screen.getByText(submission.text);
     expect(submissionText).toBeInTheDocument();
   });
 
   it("should display the submission status", () => {
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
     const submissionDate = screen.getByText(/submissions.submitted/);
     expect(submissionDate).toBeInTheDocument();
   });
 
   it("should display evaluation points", () => {
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
-    const evaluationPoints = screen.getByText(submission().metadata.evaluation.points.toString());
+    const evaluationPoints = screen.getByText(submission.metadata.evaluation.points.toString());
     expect(evaluationPoints).toBeInTheDocument();
   });
 
   it("should have a link to the submission details", () => {
-    const mockSubmission = submission();
+    const mockSubmission = submission;
     const communitySlug = "test-community-slug";
     const submissionId = mockSubmission.id;
-    renderWithRedux(<SubmissionCard  submission={submission()} />);
     const submissionLink = screen.getByRole("link");
     expect(submissionLink).toBeInTheDocument();
     expect(submissionLink).toHaveAttribute("href", `/communities/${communitySlug}/challenges/test-challenge-id/submissions/${submissionId}`);
